transcript_api/requirements.txt
fastapi
uvicorn
openai-whisper
ffmpeg-python
pydantic
python-multipart